refactor: Remove shared module and inline exceptions/util into api-server
