+++ BE/api-server/build.gradle
@@ -24,7 +24,6 @@ repositories {
 }
 
 dependencies {
-    implementation project(':shared')
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
