+++ BE/api-server/src/main/java/io/github/columnwise/shortlink/domain/exception/UrlNotFoundException.java
@@ -1,7 +1,11 @@
 package io.github.columnwise.shortlink.domain.exception;
 
+/**
+ * 단축 URL 미존재 예외
+ */
 public class UrlNotFoundException extends RuntimeException {
-	public UrlNotFoundException(String message) {
-		super(message);
-	}
+    public UrlNotFoundException(String message) {
+        super(message);
+    }
 }
+
