+++ BE/api-server/src/main/java/io/github/columnwise/shortlink/util/HashUtils.java
@@ -42,3 +42,4 @@ public final class HashUtils {
         }
     }
 }
+
