--- BE/shared/build.gradle
@@ -1,46 +1,0 @@
-plugins {
-    id 'java-library'
-    id 'io.spring.dependency-management' version '1.1.7'
-}
-
-group = 'io.github.columnwise'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencyManagement {
-    imports {
-        mavenBom 'org.springframework.boot:spring-boot-dependencies:3.5.4'
-    }
-}
-
-dependencies {
-    api 'org.springframework:spring-context'
-    api 'org.springframework.data:spring-data-jpa'
-    api 'org.springframework.data:spring-data-redis'
-    api 'jakarta.persistence:jakarta.persistence-api'
-    api 'jakarta.validation:jakarta.validation-api'
-    compileOnly 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
